% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.softermax.R
\name{as_tibble.softermax}
\alias{as_tibble.softermax}
\alias{as_tibble.softermaxExperiment}
\alias{as_tibble.softermaxPlate}
\alias{as_tibble.softermaxWavelength}
\alias{as_tibble.softermaxWell}
\title{Coerce SoftMax Pro data into tibbles}
\usage{
\method{as_tibble}{softermax}(x, nest = FALSE, ...)

\method{as_tibble}{softermaxExperiment}(x, nest = FALSE, ...)

\method{as_tibble}{softermaxPlate}(x, nest = FALSE, ...)

\method{as_tibble}{softermaxWavelength}(x, nest = FALSE, ...)

\method{as_tibble}{softermaxWell}(x, nest = FALSE, ...)
}
\arguments{
\item{x}{A \code{softermax} object}

\item{nest}{Logical value indicating whether or not the resulting tibble
should be nested (default: \code{FALSE}).}

\item{...}{Additional arguments passed to \code{\link[base]{as.data.frame}}}
}
\value{
A \code{tbl_df} object (see: \code{\link[tibble]{tibble-package}})
}
\description{
Coerce SoftMax Pro data into tibbles
}
\note{
Nesting could be done in different ways, for example by Time in
kinetic reads. For these cases, nest manually with \code{\link[tidyr]{nest}}.
}
\examples{
\dontrun{
d <- read_softmax_xml("mydata")
dX <- as_tibble(d, nest = TRUE)
}
}
