% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plates.R
\name{combine_plates}
\alias{combine_plates}
\title{Combine Data from Multiple Plates into one Data Frame}
\usage{
combine_plates(s)
}
\arguments{
\item{s}{A \code{softermax} object}
}
\value{
A data frame
}
\description{
Combine Data from Multiple Plates into one Data Frame
}
\examples{
# TODO
}
