% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.softermax.R
\name{is.softermax}
\alias{is.softermax}
\alias{is.softermax.experiment}
\alias{is.softermax.plate}
\alias{is.softermax.wavelength}
\alias{is.softermax.well}
\alias{is.softermax.note}
\alias{is.softermax.template}
\title{Test softermax Objects}
\usage{
is.softermax(x)

is.softermax.experiment(x)

is.softermax.plate(x)

is.softermax.wavelength(x)

is.softermax.well(x)

is.softermax.note(x)

is.softermax.template(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value indicating whether (\code{TRUE}) or not
(\code{FALSE}) the object inherits from the corresponding \code{softermax}
class.
}
\description{
\code{is.softermax} tests if an object is a softermax object

\code{is.softermax.experiment} tests if an object is a
softermax.experiment object

\code{is.softermax.plate} tests if an object is a
softermax.plate object

\code{is.softermax.wavelength} tests if an object is a
softermax.wavelength object

\code{is.softermax.well} tests if an object is a softermax.well
object

\code{is.softermax.note} tests if an object is a softermax.note
object

\code{is.softermax.template} tests if an object is a
softermax.template object
}
