% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_names.R
\name{note_names}
\alias{note_names}
\title{Return the names of all notes in an experiment}
\usage{
note_names(x)
}
\arguments{
\item{x}{A \code{softermax.experiment} object containing information
about an experiment in 'SoftMax Pro'}
}
\value{
A vector of strings
}
\description{
Return the names of all notes in an experiment
}
\examples{
\dontrun{
d <- read_softmax_xml("myfile.xml")
note_names(d$experiments[[1]])
}
}
