% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_names.R
\name{plate_names}
\alias{plate_names}
\title{Return the names of all plates in an experiment}
\usage{
plate_names(x)
}
\arguments{
\item{x}{A \code{softermax.experiment} object containing information
about an experiment in 'SoftMax Pro'}
}
\value{
A vector of strings
}
\description{
Return the names of all plates in an experiment
}
\examples{
\dontrun{
d <- read_softmax_xml("myfile.xml")
plate_names(d$experiments[[1]])
}
}
