% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_softmax6_template.R
\name{read_softmax6_template}
\alias{read_softmax6_template}
\title{Read a SoftMax Pro 6 Template File}
\usage{
read_softmax6_template(file, include_unspecified = FALSE,
  zeropad_wells = FALSE, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either a
single string or a raw vector).}

\item{include_unspecified}{Logical value indicating whether (\code{TRUE}) or
not (\code{FALSE}) to include non-annotated wells (default: \code{FALSE})}

\item{zeropad_wells}{Logical value indicating whether or not well names should
be zero padded (e.g., "A04") (default: \code{FALSE})}

\item{...}{Additional parameters (not used)}
}
\value{
A data frame
}
\description{
Read a SoftMax Pro 6 Template File
}
\examples{
\dontrun{
t1 <- read_softmax6_template("plate_template.txt")
}
}
