% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_softmax_text.R
\name{read_softmax_text}
\alias{read_softmax_text}
\title{Read a SoftMax Pro Text File}
\usage{
read_softmax_text(file, ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either a
single string or a raw vector).}

\item{...}{Additional arguments (not used)}
}
\value{
A \code{softermax} object that contains data for each experiment and
plate specified in the given file
}
\description{
Read a SoftMax Pro Text File
}
\examples{
\dontrun{
d <- read_softmax_text("myfile.txt")
}

}
