% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softermax-package.R
\docType{package}
\name{softermax-package}
\alias{softermax-package}
\alias{_PACKAGE}
\title{softermax: Read Exported Data from 'SoftMax Pro'}
\description{
Read microtiter plate data and templates exported from Molecular
Devices 'SoftMax Pro' software <https://www.moleculardevices.com/systems/microplate-readers/softmax-pro-7-software>.
Data exported by 'SoftMax Pro' version 5.4 and greater are supported.
}
\note{
This project and its author are not affiliated with Molecular Devices, LLC.
}
\seealso{
\url{https://www.moleculardevices.com/systems/microplate-readers/softmax-pro-7-software}
}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net}

}
