% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.softermax.R
\name{as.data.frame.softermax}
\alias{as.data.frame.softermax}
\alias{as.data.frame.softermaxExperiment}
\alias{as.data.frame.softermaxPlate}
\alias{as.data.frame.softermaxWavelength}
\alias{as.data.frame.softermaxWell}
\title{Coerce SoftMax Pro data into data frames}
\usage{
\method{as.data.frame}{softermax}(x, experimentsAsFactors = TRUE,
  platesAsFactors = TRUE, wellsAsFactors = TRUE, ...)

\method{as.data.frame}{softermaxExperiment}(x, experimentsAsFactors = TRUE,
  platesAsFactors = TRUE, wellsAsFactors = TRUE, ...)

\method{as.data.frame}{softermaxPlate}(x, platesAsFactors = TRUE,
  wellsAsFactors = TRUE, ...)

\method{as.data.frame}{softermaxWavelength}(x, wellsAsFactors = TRUE, ...)

\method{as.data.frame}{softermaxWell}(x, wellsAsFactors = TRUE, ...)
}
\arguments{
\item{x}{A \code{softermaxPlate} object}

\item{experimentsAsFactors}{Logical value indicating whether or not
experiment names should be treated as factors (default: \code{TRUE})}

\item{platesAsFactors}{Logical value indicating whether or not plate names
should be treated as factors (default: \code{TRUE})}

\item{wellsAsFactors}{Logical value indicating whether or not well names
(e.g., "H5") should be treated as factors (default: \code{TRUE})}

\item{...}{Additional parameters (not used)}
}
\value{
A data frame
}
\description{
Coerce SoftMax Pro data into data frames
}
\examples{
\dontrun{
d <- as.data.frame(read_softmax_xml("myfile.xml"))
}

}
