% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.softermax.R
\name{as.data.frame.softermax}
\alias{as.data.frame.softermax}
\alias{as.data.frame.softermaxExperiment}
\alias{as.data.frame.softermaxPlate}
\alias{as.data.frame.softermaxWavelength}
\alias{as.data.frame.softermaxWell}
\title{TODO}
\usage{
\method{as.data.frame}{softermax}(x, experimentsAsFactors = TRUE,
  platesAsFactors = TRUE, wellsAsFactors = TRUE, ...)

\method{as.data.frame}{softermaxExperiment}(x, experimentsAsFactors = TRUE,
  platesAsFactors = TRUE, wellsAsFactors = TRUE, ...)

\method{as.data.frame}{softermaxPlate}(x, platesAsFactors = TRUE,
  wellsAsFactors = TRUE, ...)

\method{as.data.frame}{softermaxWavelength}(x, wellsAsFactors = TRUE, ...)

\method{as.data.frame}{softermaxWell}(x, wellsAsFactors = TRUE, ...)
}
\arguments{
\item{x}{A \code{softermaxPlate} object}

\item{experimentsAsFactors}{TODO}

\item{platesAsFactors}{TODO}

\item{wellsAsFactors}{TODO}

\item{...}{TODO}
}
\value{
A data frame
}
\description{
TODO
}
\examples{
\dontrun{
TODO
}

}
