% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_softmax_xml.R
\name{read_softmax_xml}
\alias{read_softmax_xml}
\title{Read a SoftMax Pro XML File}
\usage{
read_softmax_xml(file, experimentsAsFactors = TRUE, platesAsFactors = TRUE,
  wellsAsFactors = TRUE)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data (either a single string or a raw vector).}

\item{experimentsAsFactors}{Logical value indicating whether or not experiment names should be factors (default: \code{TRUE})}

\item{platesAsFactors}{Logical value indicating whether or not plate names should be factors (default: \code{TRUE})}

\item{wellsAsFactors}{Logical value indicating whether or not well labels should be factors (default: \code{TRUE})}
}
\value{
TODO
}
\description{
Read a SoftMax Pro XML File
}
\examples{
\dontrun{
d <- read_softmax_xml("myfile.xml")
}

}
