% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softermax.R
\docType{package}
\name{softermax-package}
\alias{softermax}
\alias{softermax-package}
\title{softermax: Read Exported Data from SoftMax Pro}
\description{
Read microtiter plate data exported from Molecular Devices SoftMax
Pro.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/briandconnelly/softermax}
  \item Report bugs at \url{https://github.com/briandconnelly/softermax/issues}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net}

}
